+++ lib/models/webhook_model.dart
@@ -1,4 +1,5 @@
 import 'package:botika_va/utils/util.dart';
+import 'package:remove_emoji/remove_emoji.dart';
 
 class WebHookModel {
   String? id;
@@ -115,6 +116,7 @@ class MessageModel {
 
     if (type == "text") {
       value = removeHtmlTagFromString("$value");
+
       if ((value as String).contains("**id**")) {
         value = (value as String).replaceAll("**id**", "");
         lang = "id";
@@ -148,7 +150,7 @@ class MessageModel {
       return [];
     }
 
-    return splitTextAtNearestDot(value ?? "");
+    return splitTextAtNearestDot(RemoveEmoji().clean("$value"));
   }
 
   List<String> splitTextAtNearestDot(String text, {int? chunkSize = 10}) {